Move Anniversary classes to package anniversary instead of reminder
