+++ src/main/kotlin/nl/leonjoosse/anniversaryreminder/anniversary/AnniversaryResponse.kt
@@ -1,4 +1,4 @@
-package nl.leonjoosse.anniversaryreminder.reminder
+package nl.leonjoosse.anniversaryreminder.anniversary
 
 internal data class AnniversaryResponse(val response: Any = Any(),
                                         val error: String = "")
